requirements.txt
python-telegram-bot==11.1.0
requests==2.22.0
selenium==3.141.0
six==1.12.0
urllib3==1.25.2